Move from MarketDocuments to DocumentWriters
