+++ source/OutgoingMessages.Application/DocumentWriters/NotifyWholesaleServices/NotifyWholesaleServicesXmlDocumentWriter.cs
@@ -21,7 +21,7 @@ using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Application.DocumentWriters.Xml;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
 
-namespace Energinet.DataHub.EDI.OutgoingMessages.Application.MarketDocuments.NotifyWholesaleServices;
+namespace Energinet.DataHub.EDI.OutgoingMessages.Application.DocumentWriters.NotifyWholesaleServices;
 
 public class NotifyWholesaleServicesXmlDocumentWriter : DocumentWriter
 {
