+++ source/OutgoingMessages.Application/DocumentWriters/RejectRequestAggregatedMeasureData/RejectRequestAggregatedMeasureDataXmlDocumentWriter.cs
@@ -20,16 +20,16 @@ using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Application.DocumentWriters.Xml;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
 
-namespace Energinet.DataHub.EDI.OutgoingMessages.Application.MarketDocuments.RejectRequestWholesaleSettlement;
+namespace Energinet.DataHub.EDI.OutgoingMessages.Application.DocumentWriters.RejectRequestAggregatedMeasureData;
 
-public class RejectRequestWholesaleSettlementXmlDocumentWriter : DocumentWriter
+public class RejectRequestAggregatedMeasureDataXmlDocumentWriter : DocumentWriter
 {
-    public RejectRequestWholesaleSettlementXmlDocumentWriter(IMessageRecordParser parser)
+    public RejectRequestAggregatedMeasureDataXmlDocumentWriter(IMessageRecordParser parser)
         : base(
             new DocumentDetails(
-                "RejectRequestWholesaleSettlement_MarketDocument",
-                "urn:ediel.org:measure:rejectrequestwholesalesettlement:0:1 urn-ediel-org-measure-rejectrequestwholesalesettlement-0-1.xsd",
-                "urn:ediel.org:measure:rejectrequestwholesalesettlement:0:1",
+                "RejectRequestAggregatedMeasureData_MarketDocument",
+                "urn:ediel.org:measure:rejectrequestaggregatedmeasuredata:0:1 urn-ediel-org-measure-rejectrequestaggregatedmeasuredata-0-1.xsd",
+                "urn:ediel.org:measure:rejectrequestaggregatedmeasuredata:0:1",
                 "cim",
                 "ERR"),
             parser,
@@ -40,39 +40,30 @@ public class RejectRequestWholesaleSettlementXmlDocumentWriter : DocumentWriter
     public override bool HandlesType(DocumentType documentType)
     {
         ArgumentNullException.ThrowIfNull(documentType);
-        return DocumentType.RejectRequestWholesaleSettlement == documentType;
+        return DocumentType.RejectRequestAggregatedMeasureData == documentType;
     }
 
-    protected override async Task WriteMarketActivityRecordsAsync(
-        IReadOnlyCollection<string> marketActivityPayloads,
-        XmlWriter writer)
+    protected override async Task WriteMarketActivityRecordsAsync(IReadOnlyCollection<string> marketActivityPayloads, XmlWriter writer)
     {
         ArgumentNullException.ThrowIfNull(marketActivityPayloads);
         ArgumentNullException.ThrowIfNull(writer);
 
-        foreach (var wholesaleServicesRecord in ParseFrom<RejectedWholesaleServicesRecord>(marketActivityPayloads))
+        foreach (var rejectedTimeSerie in ParseFrom<RejectedTimeSerieMarketActivityRecord>(marketActivityPayloads))
         {
             await writer.WriteStartElementAsync(DocumentDetails.Prefix, "Series", null).ConfigureAwait(false);
+            await writer.WriteElementStringAsync(DocumentDetails.Prefix, "mRID", null, rejectedTimeSerie.TransactionId.ToString())
+                 .ConfigureAwait(false);
             await writer.WriteElementStringAsync(
-                    DocumentDetails.Prefix,
-                    "mRID",
-                    null,
-                    wholesaleServicesRecord.TransactionId.ToString())
-                .ConfigureAwait(false);
-            await writer.WriteElementStringAsync(
-                    DocumentDetails.Prefix,
-                    "originalTransactionIDReference_Series.mRID",
-                    null,
-                    wholesaleServicesRecord.OriginalTransactionIdReference)
-                .ConfigureAwait(false);
+                DocumentDetails.Prefix,
+                "originalTransactionIDReference_Series.mRID",
+                null,
+                rejectedTimeSerie.OriginalTransactionIdReference).ConfigureAwait(false);
 
-            foreach (var reason in wholesaleServicesRecord.RejectReasons)
+            foreach (var reason in rejectedTimeSerie.RejectReasons)
             {
                 await writer.WriteStartElementAsync(DocumentDetails.Prefix, "Reason", null).ConfigureAwait(false);
-                await writer.WriteElementStringAsync(DocumentDetails.Prefix, "code", null, reason.ErrorCode)
-                    .ConfigureAwait(false);
-                await writer.WriteElementStringAsync(DocumentDetails.Prefix, "text", null, reason.ErrorMessage)
-                    .ConfigureAwait(false);
+                await writer.WriteElementStringAsync(DocumentDetails.Prefix, "code", null, reason.ErrorCode).ConfigureAwait(false);
+                await writer.WriteElementStringAsync(DocumentDetails.Prefix, "text", null, reason.ErrorMessage).ConfigureAwait(false);
                 await writer.WriteEndElementAsync().ConfigureAwait(false);
             }
