+++ source/OutgoingMessages.Application/DocumentWriters/RejectRequestWholesaleSettlement/RejectRequestWholesaleSettlementJsonDocumentWriter.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
@@ -23,15 +24,15 @@ using Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.OutgoingMessages;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.OutgoingMessages.Queueing;
 
-namespace Energinet.DataHub.EDI.OutgoingMessages.Application.MarketDocuments.RejectRequestAggregatedMeasureData;
+namespace Energinet.DataHub.EDI.OutgoingMessages.Application.DocumentWriters.RejectRequestWholesaleSettlement;
 
-public class RejectRequestAggregatedMeasureDataJsonDocumentWriter : IDocumentWriter
+public class RejectRequestWholesaleSettlementJsonDocumentWriter : IDocumentWriter
 {
-    private const string DocumentTypeName = "RejectRequestAggregatedMeasureData_MarketDocument";
+    private const string DocumentTypeName = "RejectRequestWholesaleSettlement_MarketDocument";
     private const string TypeCode = "ERR";
     private readonly IMessageRecordParser _parser;
 
-    public RejectRequestAggregatedMeasureDataJsonDocumentWriter(IMessageRecordParser parser)
+    public RejectRequestWholesaleSettlementJsonDocumentWriter(IMessageRecordParser parser)
     {
         _parser = parser;
     }
@@ -45,13 +46,15 @@ public class RejectRequestAggregatedMeasureDataJsonDocumentWriter : IDocumentWri
 
     public bool HandlesType(DocumentType documentType)
     {
-        return documentType == DocumentType.RejectRequestAggregatedMeasureData;
+        return documentType == DocumentType.RejectRequestWholesaleSettlement;
     }
 
-    public async Task<MarketDocumentStream> WriteAsync(OutgoingMessageHeader header, IReadOnlyCollection<string> marketActivityRecords)
+    public async Task<MarketDocumentStream> WriteAsync(
+        OutgoingMessageHeader header,
+        IReadOnlyCollection<string> marketActivityRecords)
     {
         var stream = new MarketDocumentWriterMemoryStream();
-        var options = new JsonWriterOptions() { Indented = true };
+        var options = new JsonWriterOptions { Indented = true };
         using var writer = new Utf8JsonWriter(stream, options);
 
         JsonHeaderWriter.Write(header, DocumentTypeName, TypeCode, ReasonCode.FullyRejected.Code, writer);
@@ -102,15 +105,13 @@ public class RejectRequestAggregatedMeasureDataJsonDocumentWriter : IDocumentWri
         writer.WriteEndObject();
     }
 
-    private ReadOnlyCollection<RejectedTimeSerieMarketActivityRecord> ParseFrom(IReadOnlyCollection<string> payloads)
+    private ReadOnlyCollection<RejectedWholesaleServicesRecord> ParseFrom(IReadOnlyCollection<string> payloads)
     {
         ArgumentNullException.ThrowIfNull(payloads);
-        var timeSeries = new List<RejectedTimeSerieMarketActivityRecord>();
-        foreach (var payload in payloads)
-        {
-            timeSeries.Add(_parser.From<RejectedTimeSerieMarketActivityRecord>(payload));
-        }
 
-        return timeSeries.AsReadOnly();
+        return payloads
+            .Select(payload => _parser.From<RejectedWholesaleServicesRecord>(payload))
+            .ToList()
+            .AsReadOnly();
     }
 }
